*: merge avalanched lite to avalanche
