--- bin/avalanched-aws-lite/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-name = "avalanched-aws-lite"
-version = "0.0.1"
-edition = "2021"
-rust-version = "1.62"
-
-[dependencies]
-avalanche-installer = "0.0.2"
-avalanche-sdk = { version = "0.0.16", features = ["cloudwatch"] } # https://crates.io/crates/avalanche-sdk
-avalanche-types = { version = "0.0.23", features = ["jsonrpc", "metrics"] } # https://crates.io/crates/avalanche-types
-avalanchego = { path = "../../crates/avalanchego" }
-avalancheup-aws = { path = "../../bin/avalancheup-aws" }
-aws-manager = "0.0.9"
-aws-sdk-cloudwatch = "0.15.0"
-aws-sdk-ec2 = "0.15.0"
-aws-sdk-s3 = "0.15.0"
-clap = { version = "3.2.12", features = ["cargo", "derive"] } # https://github.com/clap-rs/clap/releases
-command-manager = "0.0.1"
-compress-manager = "0.0.2"
-coreth = { path = "../../crates/coreth" }
-env_logger = "0.9.0"
-infra-aws = { path = "../../infra-aws" }
-log = "0.4.17"
-random-manager = "0.0.1"
-tempfile = "3.3.0"
-tokio = { version = "1.20.0", features = ["full"] } # https://github.com/tokio-rs/tokio/releases
