--- bin/avalanched-aws-lite/src/flags.rs
@@ -1,5 +1,0 @@
-/// Defines flag options.
-pub struct Options {
-    pub log_level: String,
-    pub lite_mode: bool,
-}
