+++ bin/avalanched-aws/Cargo.toml
@@ -1,17 +1,12 @@
 [package]
 name = "avalanched-aws"
-version = "0.0.10"
+version = "0.0.1"
 edition = "2021"
 rust-version = "1.62"
-
-[[bin]]
-name = "avalanched-aws"
-path = "src/main.rs"
 
 [dependencies]
 avalanche-installer = "0.0.2"
 avalanche-sdk = { version = "0.0.16", features = ["cloudwatch"] } # https://crates.io/crates/avalanche-sdk
-coreth = { path = "../../crates/coreth" }
 avalanche-types = { version = "0.0.23", features = ["jsonrpc", "metrics"] } # https://crates.io/crates/avalanche-types
 avalanchego = { path = "../../crates/avalanchego" }
 avalancheup-aws = { path = "../../bin/avalancheup-aws" }
@@ -22,6 +17,7 @@ aws-sdk-s3 = "0.15.0"
 clap = { version = "3.2.12", features = ["cargo", "derive"] } # https://github.com/clap-rs/clap/releases
 command-manager = "0.0.1"
 compress-manager = "0.0.2"
+coreth = { path = "../../crates/coreth" }
 env_logger = "0.9.0"
 infra-aws = { path = "../../infra-aws" }
 log = "0.4.17"
