+++ bin/avalanched-aws/src/command.rs
@@ -4,7 +4,6 @@ use std::{
     os::unix::fs::PermissionsExt,
     path::Path,
     sync::Arc,
-    time::SystemTime,
 };
 
 use crate::flags;
@@ -16,7 +15,6 @@ use aws_manager::{
     s3,
 };
 use aws_sdk_ec2::model::{Filter, Tag};
-use aws_sdk_s3::model::Object;
 use infra_aws::{certs, telemetry};
 use tokio::time::{sleep, Duration};
 
@@ -30,35 +28,58 @@ pub async fn execute(opts: flags::Options) -> io::Result<()> {
     let meta = fetch_metadata().await?;
 
     let aws_creds = load_aws_credential(&meta.region).await?;
+    let ec2_manager_arc = Arc::new(aws_creds.ec2_manager.clone());
     let s3_manager_arc = Arc::new(aws_creds.s3_manager.clone());
+    let cloudwatch_manager_arc = Arc::new(aws_creds.cw_manager.clone());
 
     let tags = fetch_tags(
-        Arc::new(aws_creds.ec2_manager.clone()),
+        Arc::clone(&ec2_manager_arc),
         Arc::new(meta.ec2_instance_id.clone()),
     )
     .await?;
 
-    install_avalanche(
-        Arc::clone(&s3_manager_arc),
-        &tags.s3_bucket,
-        &tags.id,
-        &tags.avalanche_bin_path,
-    )
-    .await?;
+    if opts.lite_mode {
+        install_avalanche_from_github(
+            &tags.os_type.clone(),
+            &tags.arch_type.clone(),
+            &tags.avalanche_bin_path.clone(),
+        )
+        .await?;
+    } else {
+        install_avalanche_from_s3(
+            Arc::clone(&s3_manager_arc),
+            &tags.s3_bucket,
+            &tags.id,
+            &tags.avalanche_bin_path.clone(),
+        )
+        .await?;
+    }
 
-    let mut spec = load_spec(
-        Arc::clone(&s3_manager_arc),
-        &tags.s3_bucket,
-        &tags.id,
-        &meta.public_ipv4,
-        &tags.avalancheup_spec_path,
-    )
-    .await?;
+    let (mut avalanchego_config, coreth_config) = if opts.lite_mode {
+        let avalanchego_config =
+            write_default_avalanche_config(tags.network_id, &meta.public_ipv4)?;
+        let coreth_config =
+            write_default_coreth_config(&avalanchego_config.chain_config_dir.clone())?;
+
+        (avalanchego_config, coreth_config)
+    } else {
+        let spec = download_spec(
+            Arc::clone(&s3_manager_arc),
+            &tags.s3_bucket,
+            &tags.id,
+            &meta.public_ipv4,
+            &tags.avalancheup_spec_path,
+        )
+        .await?;
+        write_coreth_config_from_spec(&spec)?;
+
+        (spec.avalanchego_config.clone(), spec.coreth_config.clone())
+    };
 
     create_cloudwatch_config(
         &tags.id,
         tags.node_kind.clone(),
-        &spec.avalanchego_config.log_dir,
+        &avalanchego_config.log_dir,
         true,
         true,
         &tags.avalanche_data_volume_path,
@@ -73,19 +94,11 @@ pub async fn execute(opts: flags::Options) -> io::Result<()> {
     );
     let certs_manager = certs::Manager {
         envelope_manager,
-        s3_manager: aws_creds.clone().s3_manager.clone(),
+        s3_manager: aws_creds.s3_manager.clone(),
         s3_bucket: tags.s3_bucket.to_string(),
     };
-    let tls_key_path = spec
-        .avalanchego_config
-        .staking_tls_key_file
-        .clone()
-        .unwrap();
-    let tls_cert_path = spec
-        .avalanchego_config
-        .staking_tls_cert_file
-        .clone()
-        .unwrap();
+    let tls_key_path = avalanchego_config.staking_tls_key_file.clone().unwrap();
+    let tls_cert_path = avalanchego_config.staking_tls_cert_file.clone().unwrap();
     let (node_id, newly_generated) = certs_manager
         .load_or_generate(&tls_key_path, &tls_cert_path)
         .await?;
@@ -98,16 +111,8 @@ pub async fn execute(opts: flags::Options) -> io::Result<()> {
     if newly_generated {
         log::info!("STEP: backing up newly generated certificates...");
 
-        let s3_key_tls_key = format!(
-            "{}/{}.key.zstd.encrypted",
-            avalancheup_aws::StorageNamespace::PkiKeyDir(tags.id.clone()).encode(),
-            node_id
-        );
-        let s3_key_tls_cert = format!(
-            "{}/{}.crt.zstd.encrypted",
-            avalancheup_aws::StorageNamespace::PkiKeyDir(tags.id.clone()).encode(),
-            node_id
-        );
+        let s3_key_tls_key = format!("{}/pki/{}.key.zstd.encrypted", tags.id, node_id);
+        let s3_key_tls_cert = format!("{}/pki/{}.crt.zstd.encrypted", tags.id, node_id);
 
         certs_manager
             .upload(
@@ -154,15 +159,21 @@ pub async fn execute(opts: flags::Options) -> io::Result<()> {
                 .set_values(Some(vec![tags.id.clone()]))
                 .build(),
         ];
+
         let volumes = aws_creds
             .ec2_manager
             .describe_volumes(Some(filters))
             .await
-            .expect("failed describe_volumes");
+            .map_err(|e| Error::new(ErrorKind::Other, format!("failed describe_volumes {}", e)))?;
         log::info!("found {} attached volume", volumes.len());
-        assert!(volumes.len() == 1);
-        let volume_id = volumes[0].volume_id().unwrap();
+        if volumes.len() != 1 {
+            return Err(Error::new(
+                ErrorKind::Other,
+                format!("unexpected {} volumes found", volumes.len()),
+            ));
+        }
 
+        let volume_id = volumes[0].volume_id().unwrap();
         if let Some(v) = volumes[0].tags() {
             let mut tag_exists = false;
             for tg in v.iter() {
@@ -206,10 +217,86 @@ pub async fn execute(opts: flags::Options) -> io::Result<()> {
         }
     }
 
+    if !opts.lite_mode
+        && avalanchego_config.is_custom_network()
+        && avalanchego_config.genesis.is_some()
+        && !Path::new(&avalanchego_config.clone().genesis.unwrap()).exists()
+    // check "exists" for idempotency
+    {
+        log::info!("STEP: running discover/genesis updates for custom network...");
+
+        if matches!(tags.node_kind, node::Kind::Anchor) {
+            let bootstrappiong_anchor_node_s3_keys = discover_bootstrapping_anchor_nodes_from_s3(
+                &meta.ec2_instance_id,
+                &node_id.to_string(),
+                &meta.public_ipv4,
+                Arc::clone(&s3_manager_arc),
+                &tags.s3_bucket,
+                &tags.avalancheup_spec_path,
+            )
+            .await?;
+
+            merge_bootstrapping_anchor_nodes_to_write_genesis(
+                bootstrappiong_anchor_node_s3_keys,
+                &tags.avalancheup_spec_path,
+            )?;
+
+            // for now, just overwrite from every seed anchor node
+            sleep(Duration::from_secs(1)).await;
+
+            log::info!("STEP: uploading the new genesis file from each anchor node, which is to be shared with non-anchor nodes");
+            s3::spawn_put_object(
+                aws_creds.s3_manager.clone(),
+                &avalanchego_config.clone().genesis.unwrap(),
+                &tags.s3_bucket,
+                &avalancheup_aws::StorageNamespace::GenesisFile(tags.id.clone()).encode(),
+            )
+            .await
+            .map_err(|e| Error::new(ErrorKind::Other, format!("failed spawn_put_object {}", e)))?;
+        }
+
+        if matches!(tags.node_kind, node::Kind::NonAnchor) {
+            download_genesis_from_ready_anchor_nodes(
+                Arc::clone(&s3_manager_arc),
+                &tags.s3_bucket,
+                &tags.avalancheup_spec_path,
+            )
+            .await?;
+
+            let (bootstrap_ids, bootstrap_ips) = discover_ready_anchor_nodes_from_s3(
+                Arc::clone(&s3_manager_arc),
+                &tags.s3_bucket,
+                &tags.avalancheup_spec_path,
+            )
+            .await?;
+
+            // whenever updating "avalanchego_config", we must persist locally
+            // via spec file for next spec loads
+            log::info!(
+                "updating spec and avalanchego_config with bootstrap Ids '{:?}' and Ips '{:?}'",
+                bootstrap_ids,
+                bootstrap_ips
+            );
+            avalanchego_config.bootstrap_ids = Some(bootstrap_ids.join(","));
+            avalanchego_config.bootstrap_ips = Some(bootstrap_ips.join(","));
+
+            let mut spec = avalancheup_aws::Spec::load(&tags.avalancheup_spec_path)?;
+            spec.avalanchego_config = avalanchego_config.clone();
+
+            spec.sync(&tags.avalancheup_spec_path)?;
+            spec.avalanchego_config.sync(None)?;
+        }
+    }
+
+    start_avalanche_systemd_service(
+        &tags.avalanche_bin_path,
+        &avalanchego_config,
+        &coreth_config,
+    )?;
+
     let http_scheme = {
-        if spec.avalanchego_config.http_tls_enabled.is_some()
-            && spec
-                .avalanchego_config
+        if avalanchego_config.http_tls_enabled.is_some()
+            && avalanchego_config
                 .http_tls_enabled
                 .expect("unexpected None avalanchego_config.http_tls_enabled")
         {
@@ -218,523 +305,45 @@ pub async fn execute(opts: flags::Options) -> io::Result<()> {
             "http"
         }
     };
-    let local_node = avalancheup_aws::Node::new(
-        tags.clone().node_kind.clone(),
-        &meta.ec2_instance_id,
-        &node_id.to_string(),
-        &meta.public_ipv4,
-        http_scheme,
-        spec.avalanchego_config.http_port,
-    );
-    log::info!(
-        "loaded node:\n{}",
-        local_node
-            .encode_yaml()
-            .expect("failed to encode node Info")
-    );
-
-    if spec.avalanchego_config.is_custom_network()
-        && matches!(tags.node_kind, node::Kind::Anchor)
-        && spec.avalanchego_config.genesis.is_some()
-        && !Path::new(&spec.avalanchego_config.clone().genesis.unwrap()).exists()
-    {
-        log::info!("STEP: publishing seed/bootstrapping anchor node information for discovery");
-        let node_info = avalancheup_aws::NodeInfo::new(
-            local_node.clone(),
-            spec.avalanchego_config.clone(),
-            spec.coreth_config.clone(),
-        );
-        let tmp_path =
-            random_manager::tmp_path(10, Some(".yaml")).expect("unexpected tmp_path failure");
-        node_info.sync(tmp_path.clone()).unwrap();
-
-        s3::spawn_put_object(
-            aws_creds.s3_manager.clone(),
-            &tmp_path,
-            &tags.s3_bucket,
-            &avalancheup_aws::StorageNamespace::DiscoverBootstrappingAnchorNode(
-                tags.id.clone(),
-                local_node.clone(),
-            )
-            .encode(),
-        )
-        .await
-        .expect("failed s3::spawn_put_object");
-
-        fs::remove_file(tmp_path).expect("failed fs::remove_file");
-
-        sleep(Duration::from_secs(30)).await;
-        log::info!("STEP: waiting for all seed/bootstrapping anchor nodes to be ready");
-        let target_nodes = spec.machine.anchor_nodes.unwrap();
-        let s3_key = s3::append_slash(
-            &avalancheup_aws::StorageNamespace::DiscoverBootstrappingAnchorNodesDir(
-                tags.id.clone(),
-            )
-            .encode(),
-        );
-        let mut objects: Vec<Object>;
-        loop {
-            sleep(Duration::from_secs(20)).await;
-            objects = s3::spawn_list_objects(
-                aws_creds.s3_manager.clone(),
-                &tags.s3_bucket,
-                Some(s3_key.clone()),
-            )
-            .await
-            .expect("failed s3::spawn_list_objects");
-            log::info!(
-                "{} seed/bootstrapping anchor nodes are ready (expecting {} nodes)",
-                objects.len(),
-                target_nodes
-            );
-            if objects.len() as u32 >= target_nodes {
-                break;
-            }
-        }
-
-        log::info!("STEP: update genesis file with seed/bootstrapping anchor nodes");
-        let mut initial_stakers: Vec<avalanchego_genesis::Staker> = vec![];
-
-        // "initial_staked_funds" is reserved for locked P-chain balance
-        // with "spec.generated_seed_private_key_with_locked_p_chain_balance"
-        let seed_priv_keys = spec.clone().generated_seed_private_keys.unwrap();
-        let seed_priv_key = seed_priv_keys[0].clone();
-        for obj in objects.iter() {
-            let s3_key = obj.key().expect("unexpected None s3 object");
-
-            // just parse the s3 key name
-            // to reduce "s3_manager.get_object" call volume
-            let seed_anchor_node =
-                avalancheup_aws::StorageNamespace::parse_node_from_path(s3_key).unwrap();
-
-            let mut staker = avalanchego_genesis::Staker::default();
-            staker.node_id = Some(seed_anchor_node.node_id);
-            staker.reward_address = Some(
-                seed_priv_key
-                    .addresses
-                    .get(&format!("{}", spec.avalanchego_config.network_id))
-                    .unwrap()
-                    .x_address
-                    .clone(),
-            );
-
-            initial_stakers.push(staker);
-        }
-        log::info!(
-            "found {} seed anchor nodes for initial stakers",
-            initial_stakers.len()
-        );
-
-        let avalanchego_genesis_path = spec.avalanchego_config.clone().genesis.unwrap();
-        let mut avalanchego_genesis_template = spec
-            .avalanchego_genesis_template
-            .clone()
-            .expect("unexpected None avalanchego_genesis_template for custom network");
-        avalanchego_genesis_template.initial_stakers = Some(initial_stakers);
-        avalanchego_genesis_template
-            .sync(&avalanchego_genesis_path)
-            .expect("failed to sync avalanchego_genesis_path");
-
-        // for now, just overwrite from every seed anchor node
-        sleep(Duration::from_secs(1)).await;
-
-        log::info!("STEP: upload the new genesis file, to be shared with non-anchor nodes");
-        s3::spawn_put_object(
-            aws_creds.s3_manager.clone(),
-            &avalanchego_genesis_path,
-            &tags.s3_bucket,
-            &avalancheup_aws::StorageNamespace::GenesisFile(spec.id.clone()).encode(),
-        )
-        .await
-        .expect("failed s3::spawn_put_object");
-    }
-
-    if spec.avalanchego_config.is_custom_network()
-        && matches!(tags.node_kind, node::Kind::NonAnchor)
-        && spec.avalanchego_config.genesis.is_some()
-        && !Path::new(&spec.avalanchego_config.clone().genesis.unwrap()).exists()
-    {
-        log::info!("STEP: downloading genesis file from S3 (updated from other anchor nodes)");
-        let tmp_genesis_path = random_manager::tmp_path(15, Some(".json")).unwrap();
-        s3::spawn_get_object(
-            aws_creds.s3_manager.clone(),
-            &tags.s3_bucket,
-            &avalancheup_aws::StorageNamespace::GenesisFile(spec.id.clone()).encode(),
-            &tmp_genesis_path,
-        )
-        .await
-        .expect("failed s3::spawn_get_object");
-        fs::copy(
-            &tmp_genesis_path,
-            spec.avalanchego_config.clone().genesis.unwrap(),
-        )
-        .expect("failed fs::copy genesis file");
-    }
-
-    // validate after downloading genesis file
-    spec.avalanchego_config.validate().unwrap();
-    if spec.avalanchego_config.config_file.is_none() {
-        panic!("'spec.avalanchego_config.config_file' not found")
-    }
-
-    // mainnet/other pre-defined test nets have hard-coded anchor nodes
-    // thus no need for anchor nodes
-    if spec.avalanchego_config.is_custom_network()
-        && matches!(tags.node_kind, node::Kind::NonAnchor)
-    {
-        sleep(Duration::from_secs(1)).await;
-        log::info!(
-            "STEP: downloading anchor node information for network '{}'",
-            spec.avalanchego_config.network_id,
-        );
-
-        // "avalanche-ops" should always set up anchor nodes first
-        // so here we assume anchor nodes are already set up
-        // and their information is already available via shared,
-        // remote storage for service discovery
-        // so that we block non-anchor nodes until anchor nodes are ready
-        //
-        // always send a new "list_objects" on remote storage
-        // rather than relying on potentially stale (not via "spec")
-        // in case the member lists for "anchor" nodes becomes stale
-        // (e.g., machine replacement in "anchor" nodes ASG)
-        //
-        // TODO: handle stale anchor nodes by heartbeats timestamps
-        let target_nodes = spec
-            .machine
-            .anchor_nodes
-            .expect("unexpected None machine.anchor_nodes for custom network");
-        let s3_key = s3::append_slash(
-            &avalancheup_aws::StorageNamespace::DiscoverReadyAnchorNodesDir(tags.id.clone())
-                .encode(),
-        );
-        let mut objects: Vec<Object>;
-        loop {
-            sleep(Duration::from_secs(20)).await;
-
-            objects = s3::spawn_list_objects(
-                aws_creds.s3_manager.clone(),
-                &tags.s3_bucket,
-                Some(s3_key.clone()),
-            )
-            .await
-            .expect("failed s3::spawn_list_objects");
-            log::info!(
-                "{} anchor nodes are ready (expecting {} nodes)",
-                objects.len(),
-                target_nodes
-            );
-            if objects.len() as u32 >= target_nodes {
-                break;
-            }
-        }
-
-        log::info!("STEP: updating bootstrap IPs/IDs with all anchor nodes");
-        let mut bootstrap_ips: Vec<String> = vec![];
-        let mut bootstrap_ids: Vec<String> = vec![];
-        for obj in objects.iter() {
-            let s3_key = obj.key().expect("unexpected None s3 object");
-
-            // just parse the s3 key name
-            // to reduce "s3_manager.get_object" call volume
-            let anchor_node = avalancheup_aws::StorageNamespace::parse_node_from_path(s3_key)
-                .expect("failed to parse node from storage path");
-
-            // assume all nodes in the network use the same ports
-            // ref. "avalanchego/config.StakingPortKey" default value is "9651"
-            let staking_port = spec.avalanchego_config.staking_port;
-            bootstrap_ips.push(format!("{}:{}", anchor_node.public_ip, staking_port));
-            bootstrap_ids.push(anchor_node.node_id);
-        }
-        log::info!("found {} bootstrap nodes", bootstrap_ids.len());
-
-        spec.avalanchego_config.bootstrap_ips = Some(bootstrap_ips.join(","));
-        spec.avalanchego_config.bootstrap_ids = Some(bootstrap_ids.join(","));
-    }
-
-    write_avalanche_config(tags.network_id, &spec)?;
-    write_coreth_config_from_spec(&spec)?;
-
-    if spec.avalanchego_config.subnet_config_dir.is_some() {
-        let subnet_config_dir = spec
-            .avalanchego_config
-            .clone()
-            .subnet_config_dir
-            .expect("unexpected None subnet_config_dir");
-        fs::create_dir_all(Path::new(&subnet_config_dir).join("C"))?;
-    };
-
-    // TODO: upload profile files to s3?
-    if spec.avalanchego_config.profile_dir.is_some() {
-        let profile_dir = spec
-            .avalanchego_config
-            .clone()
-            .profile_dir
-            .expect("unexpected None profile_dir");
-        fs::create_dir_all(&profile_dir).expect("failed to create profile_dir");
-    };
-
-    // TODO: upload profile files to s3?
-    if spec.coreth_config.continuous_profiler_dir.is_some() {
-        let continuous_profiler_dir = spec
-            .coreth_config
-            .clone()
-            .continuous_profiler_dir
-            .expect("unexpected None continuous_profiler_dir");
-        fs::create_dir_all(&continuous_profiler_dir)
-            .expect("failed to create continuous_profiler_dir");
-    };
-
-    start_avalanche_systemd_service(&tags.avalanche_bin_path, spec.avalanchego_config.clone())?;
-
     let ep = format!(
         "{}://{}:{}",
-        http_scheme, meta.public_ipv4, spec.avalanchego_config.http_port
+        http_scheme, meta.public_ipv4, avalanchego_config.http_port
     );
     check_liveness(&ep).await?;
 
-    log::info!("spawning async routines...");
-    let node_info_ready_s3_key = {
-        if matches!(tags.node_kind, node::Kind::Anchor) {
-            avalancheup_aws::StorageNamespace::DiscoverReadyAnchorNode(
-                tags.id.clone(),
-                local_node.clone(),
-            )
-            .encode()
-        } else {
-            avalancheup_aws::StorageNamespace::DiscoverReadyNonAnchorNode(
-                tags.id.clone(),
-                local_node.clone(),
-            )
-            .encode()
-        }
-    };
-
     let handles = vec![
         tokio::spawn(publish_node_info_ready_loop(
-            aws_creds.s3_manager.clone(),
+            Arc::new(meta.ec2_instance_id.clone()),
+            tags.node_kind.clone(),
+            Arc::new(node_id.to_string()),
+            Arc::new(meta.public_ipv4.clone()),
+            Arc::clone(&s3_manager_arc),
             Arc::new(tags.s3_bucket.clone()),
-            Arc::new(node_info_ready_s3_key),
-            Arc::new(avalancheup_aws::NodeInfo::new(
-                local_node.clone(),
-                spec.avalanchego_config.clone(),
-                spec.coreth_config.clone(),
-            )),
+            Arc::new(tags.avalancheup_spec_path.clone()),
         )),
         tokio::spawn(telemetry::metrics::avalanchego::fetch_loop(
-            Arc::new(aws_creds.cw_manager.clone()),
-            Arc::new(
-                spec.clone()
-                    .aws_resources
-                    .clone()
-                    .unwrap()
-                    .cloudwatch_avalanche_metrics_namespace
-                    .unwrap(),
-            ),
+            Arc::clone(&cloudwatch_manager_arc),
+            Arc::new(tags.id.clone()),
             Duration::from_secs(120),
             Duration::from_secs(60),
-            Arc::new(local_node.http_endpoint.clone()),
-        )),
-        tokio::spawn(check_node_update_loop(
-            aws_creds.s3_manager.clone(),
-            Arc::new(tags.s3_bucket.clone()),
-            Arc::new(tags.id.clone()),
-            Arc::new(tags.avalanche_bin_path),
+            Arc::new(ep.to_string()),
         )),
     ];
 
     log::info!("STEP: blocking on handles via JoinHandle");
     for handle in handles {
-        handle.await.expect("failed handle");
+        handle.await.map_err(|e| {
+            Error::new(
+                ErrorKind::Other,
+                format!("failed await on JoinHandle {}", e),
+            )
+        })?;
     }
 
     Ok(())
 }
 
-/// if run in anchor nodes, the uploaded file will be downloaded
-/// in bootstrapping non-anchor nodes for custom networks
-async fn publish_node_info_ready_loop(
-    s3_manager: s3::Manager,
-    s3_bucket: Arc<String>,
-    s3_key: Arc<String>,
-    node_info: Arc<avalancheup_aws::NodeInfo>,
-) {
-    log::info!("STEP: starting 'publish_node_info_ready_loop'");
-
-    loop {
-        log::info!(
-            "STEP: posting node info ready for {}",
-            node_info.local_node.kind
-        );
-
-        let tmp_path =
-            random_manager::tmp_path(10, Some(".yaml")).expect("unexpected tmp_path failure");
-        node_info.sync(tmp_path.clone()).unwrap();
-
-        s3::spawn_put_object(s3_manager.clone(), &tmp_path, &s3_bucket, &s3_key.clone())
-            .await
-            .expect("failed s3::spawn_put_object");
-
-        fs::remove_file(&tmp_path).expect("failed fs::remove_file");
-
-        log::info!("sleeping 10-min for next 'publish_node_info_ready_loop'");
-        sleep(Duration::from_secs(600)).await;
-    }
-}
-
-async fn check_node_update_loop(
-    s3_manager: s3::Manager,
-    s3_bucket: Arc<String>,
-    id: Arc<String>,
-    avalanche_bin_path: Arc<String>,
-) {
-    log::info!("STEP: starting 'check_node_update_loop'");
-
-    loop {
-        log::info!("sleeping 3-min for 'check_node_update_loop'");
-        sleep(Duration::from_secs(180)).await;
-
-        log::info!("STEP: checking update artifacts event key");
-        let objects = match s3::spawn_list_objects(
-            s3_manager.clone(),
-            s3_bucket.as_str(),
-            Some(
-                avalancheup_aws::StorageNamespace::EventsUpdateArtifactsEvent(id.to_string())
-                    .encode(),
-            ),
-        )
-        .await
-        {
-            Ok(v) => v,
-            Err(e) => {
-                log::warn!("failed s3::spawn_list_objects {}, retrying...", e);
-                continue;
-            }
-        };
-
-        if objects.is_empty() {
-            log::warn!("no event key found");
-            continue;
-        }
-
-        let obj = objects[0].clone();
-        let last_modified = obj.last_modified.unwrap();
-        let last_modified_unix = last_modified.as_secs_f64();
-
-        let now = SystemTime::now();
-        let now_unix = now
-            .duration_since(SystemTime::UNIX_EPOCH)
-            .expect("unexpected None duration_since")
-            .as_secs() as f64;
-
-        // requested for the last 6-min
-        // TODO: can this be smarter...?
-        let needs_update = (now_unix - last_modified_unix) < 360_f64;
-        log::info!(
-            "last_modified_unix {}, now_unix {} [needs update: {}]",
-            last_modified_unix,
-            now_unix,
-            needs_update
-        );
-
-        if !needs_update {
-            log::warn!(
-                "update artifacts event not found (seeing {} objects)",
-                objects.len()
-            );
-            continue;
-        }
-
-        log::info!("STEP: downloading avalanched binary from S3");
-
-        // TODO: replace "avalanched" itself?
-        // TODO: fs::copy fails with 'Os { code: 26, kind: ExecutableFileBusy, message: "Text file busy" }'
-        // can't replace the process itself...
-
-        log::info!("STEP: downloading avalanche binary from S3");
-        let tmp_avalanche_bin_compressed_path =
-            random_manager::tmp_path(15, Some(".zstd")).unwrap();
-        s3::spawn_get_object(
-                    s3_manager.clone(),
-                    s3_bucket.clone().as_ref().to_string(),
-                    avalancheup_aws::StorageNamespace::EventsUpdateArtifactsInstallDirAvalancheBinCompressed(id.to_string()).encode(),
-                      tmp_avalanche_bin_compressed_path.clone(),
-                )
-                .await
-                .expect("failed s3::spawn_get_object");
-
-        log::warn!("stopping avalanche.service before unpack...");
-        command_manager::run("sudo systemctl stop avalanche.service")
-            .expect("failed systemctl stop command");
-        log::warn!("stopped avalanche.service before unpack...");
-        sleep(Duration::from_secs(10)).await;
-
-        compress_manager::unpack_file(
-            &tmp_avalanche_bin_compressed_path,
-            avalanche_bin_path.as_str(),
-            compress_manager::Decoder::Zstd,
-        )
-        .expect("failed unpack_file avalanche_bin_compressed_path");
-
-        let f = File::open(avalanche_bin_path.as_str()).expect("failed to open avalanche_bin");
-        f.set_permissions(PermissionsExt::from_mode(0o777))
-            .expect("failed to set file permission for avalanche_bin");
-        fs::remove_file(&tmp_avalanche_bin_compressed_path).expect("failed fs::remove_file");
-
-        let plugins_dir =
-            avalanche_installer::avalanchego::get_plugins_dir(avalanche_bin_path.as_str());
-        if !Path::new(&plugins_dir).exists() {
-            log::info!("STEP: creating '{}' for plugins", plugins_dir);
-            fs::create_dir_all(plugins_dir.clone()).unwrap();
-        }
-
-        log::info!("STEP: downloading plugins from S3 (if any) to overwrite");
-        let objects = s3::spawn_list_objects(
-            s3_manager.clone(),
-            s3_bucket.clone().as_ref().to_string(),
-            Some(s3::append_slash(
-                &avalancheup_aws::StorageNamespace::EventsUpdateArtifactsInstallDirPluginsDir(
-                    id.to_string(),
-                )
-                .encode(),
-            )),
-        )
-        .await
-        .expect("failed s3::spawn_list_objects");
-
-        log::info!("listed {} plugins from S3", objects.len());
-        for obj in objects.iter() {
-            let s3_key = obj.key().expect("unexpected None s3 object").to_string();
-            let tmp_path = random_manager::tmp_path(15, None).unwrap();
-            s3::spawn_get_object(
-                s3_manager.clone(),
-                s3_bucket.clone().as_ref().to_string(),
-                s3_key.clone(),
-                tmp_path.clone(),
-            )
-            .await
-            .expect("failed s3::spawn_get_object");
-
-            let file_name = extract_filename(&s3_key);
-            let file_path = format!("{}/{}", plugins_dir, file_name);
-            compress_manager::unpack_file(&tmp_path, &file_path, compress_manager::Decoder::Zstd)
-                .unwrap();
-
-            let f = File::open(file_path).expect("failed to open plugin file");
-            f.set_permissions(PermissionsExt::from_mode(0o777))
-                .expect("failed to set file permission");
-            fs::remove_file(&tmp_path).expect("failed fs::remove_file");
-        }
-
-        // updated the avalanched itself, so sleep for cloudwatch logs and restart
-        log::warn!(
-            "artifacts have been updated... will trigger avalanched restart by panic here..."
-        );
-        sleep(Duration::from_secs(240)).await; // sleep to prevent duplicate updates
-        panic!("panic avalanched to trigger restarts via systemd service!!!")
-    }
-}
-
+#[derive(Debug, Clone)]
 struct Metadata {
     az: String,
     region: String,
@@ -926,14 +535,42 @@ async fn fetch_tags(
     Ok(fetched_tags)
 }
 
-async fn install_avalanche(
+async fn install_avalanche_from_github(
+    arch_type: &str,
+    os_type: &str,
+    avalanche_bin_path: &str,
+) -> io::Result<()> {
+    log::info!("STEP: installing Avalanche from github...");
+
+    let arch = if arch_type.to_string() == "amd64" {
+        Some(avalanche_installer::avalanchego::Arch::Amd64)
+    } else {
+        None
+    };
+    let os = if os_type.to_string() == "linux" {
+        Some(avalanche_installer::avalanchego::Os::Linux)
+    } else {
+        None
+    };
+
+    let (binary_path, _) = avalanche_installer::avalanchego::download(arch, os).await?;
+    fs::copy(&binary_path, avalanche_bin_path)?;
+
+    let plugins_dir = avalanche_installer::avalanchego::get_plugins_dir(avalanche_bin_path);
+    if !Path::new(&plugins_dir).exists() {
+        log::info!("creating '{}' directory for plugins", plugins_dir);
+        fs::create_dir_all(plugins_dir.clone())?;
+    };
+
+    Ok(())
+}
+
+async fn install_avalanche_from_s3(
     s3_manager: Arc<s3::Manager>,
     s3_bucket: &str,
     id: &str,
     avalanche_bin_path: &str,
 ) -> io::Result<()> {
-    log::info!("STEP: installing Avalanche...");
-
     let s3_manager: &s3::Manager = s3_manager.as_ref();
 
     if !Path::new(avalanche_bin_path).exists() {
@@ -941,8 +578,7 @@ async fn install_avalanche(
 
         let s3_key =
             avalancheup_aws::StorageNamespace::AvalancheBinCompressed(id.to_string()).encode();
-        let tmp_avalanche_bin_compressed_path =
-            random_manager::tmp_path(15, Some(".zstd")).unwrap();
+        let tmp_avalanche_bin_compressed_path = random_manager::tmp_path(15, Some(".zstd"))?;
 
         s3::spawn_get_object(
             s3_manager.to_owned(),
@@ -967,7 +603,7 @@ async fn install_avalanche(
     let plugins_dir = avalanche_installer::avalanchego::get_plugins_dir(&avalanche_bin_path);
     if !Path::new(&plugins_dir).exists() {
         log::info!("STEP: creating '{}' for plugins", plugins_dir);
-        fs::create_dir_all(plugins_dir.clone()).unwrap();
+        fs::create_dir_all(plugins_dir.clone())?;
 
         log::info!("STEP: downloading plugins from S3 (if any)");
         let objects = s3::spawn_list_objects(
@@ -983,7 +619,7 @@ async fn install_avalanche(
         log::info!("listed {} plugins from S3", objects.len());
         for obj in objects.iter() {
             let s3_key = obj.key().expect("unexpected None s3 object").to_string();
-            let tmp_path = random_manager::tmp_path(15, None).unwrap();
+            let tmp_path = random_manager::tmp_path(15, None)?;
 
             s3::spawn_get_object(s3_manager.to_owned(), s3_bucket, &s3_key, &tmp_path)
                 .await
@@ -1004,16 +640,16 @@ async fn install_avalanche(
     Ok(())
 }
 
-async fn load_spec(
+async fn download_spec(
     s3_manager: Arc<s3::Manager>,
     s3_bucket: &str,
     id: &str,
     public_ipv4: &str,
     avalancheup_spec_path: &str,
 ) -> io::Result<avalancheup_aws::Spec> {
-    log::info!("STEP: downloading avalanche-ops spec from S3");
+    log::info!("STEP: downloading avalancheup spec file from S3...");
 
-    let tmp_spec_file_path = random_manager::tmp_path(15, Some(".yaml")).unwrap();
+    let tmp_spec_file_path = random_manager::tmp_path(15, Some(".yaml"))?;
 
     let s3_manager: &s3::Manager = s3_manager.as_ref();
     s3::spawn_get_object(
@@ -1025,7 +661,7 @@ async fn load_spec(
     .await
     .map_err(|e| Error::new(ErrorKind::Other, format!("failed spawn_get_object {}", e)))?;
 
-    let mut spec = avalancheup_aws::Spec::load(&tmp_spec_file_path).unwrap();
+    let mut spec = avalancheup_aws::Spec::load(&tmp_spec_file_path)?;
     spec.avalanchego_config.public_ip = Some(public_ipv4.to_string());
     spec.avalanchego_config.sync(None)?;
 
@@ -1035,35 +671,57 @@ async fn load_spec(
     Ok(spec)
 }
 
-fn write_avalanche_config(network_id: u32, spec: &avalancheup_aws::Spec) -> io::Result<()> {
+fn write_default_avalanche_config(
+    network_id: u32,
+    public_ipv4: &str,
+) -> io::Result<avalanchego::config::Config> {
     log::info!("STEP: writing Avalanche config file...");
 
-    if spec.avalanchego_config.network_id != network_id {
-        return Err(Error::new(
-            ErrorKind::Other,
-            format!(
-                "unexpected network Id, [tag {}, avalanchego config {}]",
-                network_id, spec.avalanchego_config.network_id
-            ),
-        ));
-    }
+    let mut avalanchego_config = match network_id {
+        1 => avalanchego::config::Config::default_main(),
+        5 => avalanchego::config::Config::default_fuji(),
+        _ => avalanchego::config::Config::default_custom(),
+    };
+    avalanchego_config.network_id = network_id;
+    avalanchego_config.public_ip = Some(public_ipv4.to_string());
+    avalanchego_config.sync(None)?;
 
-    let log_dir = spec.avalanchego_config.clone().log_dir;
-    fs::create_dir_all(&log_dir)?;
+    Ok(avalanchego_config)
+}
 
-    spec.avalanchego_config.sync(None)
+fn write_default_coreth_config(chain_config_dir: &str) -> io::Result<coreth::config::Config> {
+    log::info!("STEP: writing default coreth config file...");
+
+    fs::create_dir_all(Path::new(chain_config_dir).join("C"))?;
+
+    let default_coreth_config = coreth::config::Config::default();
+
+    let tmp_coreth_config_path = random_manager::tmp_path(15, Some(".json"))?;
+    let chain_config_c_path = Path::new(chain_config_dir).join("C").join("config.json");
+
+    default_coreth_config.sync(&tmp_coreth_config_path)?;
+    fs::copy(&tmp_coreth_config_path, &chain_config_c_path)?;
+    fs::remove_file(&tmp_coreth_config_path)?;
+
+    log::info!(
+        "saved default coreth config file to {:?}",
+        chain_config_c_path.as_os_str()
+    );
+
+    Ok(default_coreth_config)
 }
 
 fn write_coreth_config_from_spec(spec: &avalancheup_aws::Spec) -> io::Result<()> {
-    log::info!("STEP: writing coreth config file...");
+    log::info!("STEP: writing coreth config file from spec...");
 
     let chain_config_dir = spec.avalanchego_config.chain_config_dir.clone();
     fs::create_dir_all(Path::new(&chain_config_dir).join("C"))?;
 
-    let tmp_coreth_config_path = random_manager::tmp_path(15, Some(".json")).unwrap();
+    let tmp_coreth_config_path = random_manager::tmp_path(15, Some(".json"))?;
     let chain_config_c_path = Path::new(&chain_config_dir).join("C").join("config.json");
 
     spec.coreth_config.sync(&tmp_coreth_config_path)?;
+
     fs::copy(&tmp_coreth_config_path, &chain_config_c_path)?;
     fs::remove_file(&tmp_coreth_config_path)?;
 
@@ -1101,16 +759,315 @@ fn create_cloudwatch_config(
     cw_config_manager.sync(Some(vec![String::from("/var/log/avalanched.log")]))
 }
 
+/// To be called by each bootstrapping anchor node: Each anchor node publishes
+/// the anchor node information to the S3, and waits. And each anchor node
+/// lists all keys from the bootstrapping s3 directory for its peer discovery.
+/// The function returns all s3 keys that contains the anchor node information.
+async fn discover_bootstrapping_anchor_nodes_from_s3(
+    ec2_instance_id: &str,
+    node_id: &str,
+    public_ipv4: &str,
+    s3_manager: Arc<s3::Manager>,
+    s3_bucket: &str,
+    avalancheup_spec_path: &str,
+) -> io::Result<Vec<String>> {
+    log::info!("STEP: publishing bootstrapping local anchor node information for discovery...");
+
+    let spec = avalancheup_aws::Spec::load(avalancheup_spec_path)?;
+
+    let http_scheme = {
+        if spec.avalanchego_config.http_tls_enabled.is_some()
+            && spec
+                .avalanchego_config
+                .http_tls_enabled
+                .expect("unexpected None avalanchego_config.http_tls_enabled")
+        {
+            "https"
+        } else {
+            "http"
+        }
+    };
+    let local_node = avalancheup_aws::Node::new(
+        node::Kind::Anchor,
+        ec2_instance_id,
+        node_id,
+        public_ipv4,
+        http_scheme,
+        spec.avalanchego_config.http_port,
+    );
+
+    log::info!(
+        "loaded local node information:\n{}",
+        local_node
+            .encode_yaml()
+            .expect("failed to encode node Info")
+    );
+
+    let node_info = avalancheup_aws::NodeInfo::new(
+        local_node.clone(),
+        spec.avalanchego_config.clone(),
+        spec.coreth_config.clone(),
+    );
+    let node_info_path = random_manager::tmp_path(10, Some(".yaml"))?;
+    node_info.sync(node_info_path.clone()).unwrap();
+
+    let s3_manager: &s3::Manager = s3_manager.as_ref();
+    s3::spawn_put_object(
+        s3_manager.clone(),
+        node_info_path.as_str(),
+        s3_bucket,
+        &avalancheup_aws::StorageNamespace::DiscoverBootstrappingAnchorNode(
+            spec.id.clone(),
+            local_node.clone(),
+        )
+        .encode(),
+    )
+    .await
+    .map_err(|e| Error::new(ErrorKind::Other, format!("failed spawn_put_object {}", e)))?;
+
+    fs::remove_file(node_info_path)?;
+
+    log::info!("waiting some time for initial anchor nodes to start...");
+    sleep(Duration::from_secs(30)).await; // enough time for all anchor nodes up to be
+
+    let target_nodes = spec.machine.anchor_nodes.unwrap();
+    log::info!("waiting for all other seed/bootstrapping anchor nodes to publish their own (expected total {} nodes)",target_nodes);
+
+    let s3_key_prefix = s3::append_slash(
+        &avalancheup_aws::StorageNamespace::DiscoverBootstrappingAnchorNodesDir(spec.id).encode(),
+    );
+
+    let mut objects: Vec<aws_sdk_s3::model::Object>;
+    loop {
+        sleep(Duration::from_secs(20)).await;
+
+        objects =
+            s3::spawn_list_objects(s3_manager.clone(), s3_bucket, Some(s3_key_prefix.clone()))
+                .await
+                .map_err(|e| {
+                    Error::new(ErrorKind::Other, format!("failed spawn_list_objects {}", e))
+                })?;
+
+        log::info!(
+            "{} seed/bootstrapping anchor nodes are ready (expecting {} nodes)",
+            objects.len(),
+            target_nodes
+        );
+
+        if objects.len() as u32 >= target_nodes {
+            break;
+        }
+    }
+
+    let mut s3_keys: Vec<String> = Vec::new();
+    for obj in objects.iter() {
+        let s3_key = obj.key().expect("unexpected None s3 object");
+        s3_keys.push(s3_key.to_string());
+    }
+    // s3 API should return the sorted list
+    // in the descending order of "last_modified" timestamps
+    // but to make sure, sort in lexicographical order
+    s3_keys.sort();
+
+    Ok(s3_keys)
+}
+
+/// Combines all anchor node Ids and write them into a genesis for initial stakers.
+/// The genesis file path is defined in avalanchego "--genesis" flag.
+fn merge_bootstrapping_anchor_nodes_to_write_genesis(
+    bootstrapping_anchor_node_s3_keys: Vec<String>,
+    avalancheup_spec_path: &str,
+) -> io::Result<()> {
+    log::info!(
+        "STEP: combining {} anchor node S3 keys to write genesis...",
+        bootstrapping_anchor_node_s3_keys.len()
+    );
+
+    let spec = avalancheup_aws::Spec::load(avalancheup_spec_path)?;
+
+    // "initial_staked_funds" is reserved for locked P-chain balance
+    // with "spec.generated_seed_private_key_with_locked_p_chain_balance"
+    let seed_priv_keys = spec.clone().generated_seed_private_keys.unwrap();
+    let seed_priv_key = seed_priv_keys[0].clone();
+
+    let mut initial_stakers: Vec<avalanchego_genesis::Staker> = vec![];
+    for s3_key in bootstrapping_anchor_node_s3_keys.iter() {
+        // just parse the s3 key name
+        // to reduce "s3_manager.get_object" call volume
+        let seed_anchor_node = avalancheup_aws::StorageNamespace::parse_node_from_path(s3_key)?;
+
+        let mut staker = avalanchego_genesis::Staker::default();
+        staker.node_id = Some(seed_anchor_node.node_id);
+        staker.reward_address = Some(
+            seed_priv_key
+                .addresses
+                .get(&format!("{}", spec.avalanchego_config.network_id))
+                .unwrap()
+                .x_address
+                .clone(),
+        );
+
+        initial_stakers.push(staker);
+    }
+    log::info!(
+        "found {} seed anchor nodes for initial stakers",
+        initial_stakers.len()
+    );
+
+    let avalanchego_genesis_path = spec.avalanchego_config.clone().genesis.unwrap();
+
+    let mut avalanchego_genesis_template = spec
+        .avalanchego_genesis_template
+        .clone()
+        .expect("unexpected None avalanchego_genesis_template for custom network");
+    avalanchego_genesis_template.initial_stakers = Some(initial_stakers);
+    avalanchego_genesis_template.sync(&avalanchego_genesis_path)?;
+
+    Ok(())
+}
+
+async fn download_genesis_from_ready_anchor_nodes(
+    s3_manager: Arc<s3::Manager>,
+    s3_bucket: &str,
+    avalancheup_spec_path: &str,
+) -> io::Result<()> {
+    log::info!("STEP: downloading genesis file from S3 (updated from other anchor nodes)");
+
+    let spec = avalancheup_aws::Spec::load(avalancheup_spec_path)?;
+    let tmp_genesis_path = random_manager::tmp_path(15, Some(".json"))?;
+
+    let s3_manager: &s3::Manager = s3_manager.as_ref();
+    s3::spawn_get_object(
+        s3_manager.clone(),
+        s3_bucket,
+        &avalancheup_aws::StorageNamespace::GenesisFile(spec.id.clone()).encode(),
+        &tmp_genesis_path,
+    )
+    .await
+    .map_err(|e| Error::new(ErrorKind::Other, format!("failed spawn_get_object {}", e)))?;
+
+    fs::copy(
+        &tmp_genesis_path,
+        spec.avalanchego_config.clone().genesis.unwrap(),
+    )?;
+    fs::remove_file(&tmp_genesis_path)?;
+
+    Ok(())
+}
+
+/// Discover ready anchor nodes from S3 and returns the
+/// combined list of bootstrap Ids and Ips.
+/// mainnet/other pre-defined test nets have hard-coded anchor nodes
+/// thus no need for anchor nodes.
+async fn discover_ready_anchor_nodes_from_s3(
+    s3_manager: Arc<s3::Manager>,
+    s3_bucket: &str,
+    avalancheup_spec_path: &str,
+) -> io::Result<(Vec<String>, Vec<String>)> {
+    log::info!("STEP: listing S3 directory to discover ready anchor nodes...");
+
+    let spec = avalancheup_aws::Spec::load(avalancheup_spec_path)?;
+
+    // "avalanche-ops" should always set up anchor nodes first
+    // so here we assume anchor nodes are already set up
+    // and their information is already available via shared,
+    // remote storage for service discovery
+    // so that we block non-anchor nodes until anchor nodes are ready
+    //
+    // always send a new "list_objects" on remote storage
+    // rather than relying on potentially stale (not via "spec")
+    // in case the member lists for "anchor" nodes becomes stale
+    // (e.g., machine replacement in "anchor" nodes ASG)
+    //
+    // TODO: handle stale anchor nodes by heartbeats timestamps
+    let target_nodes = spec
+        .machine
+        .anchor_nodes
+        .expect("unexpected None machine.anchor_nodes for custom network");
+    let s3_key_prefix = s3::append_slash(
+        &avalancheup_aws::StorageNamespace::DiscoverReadyAnchorNodesDir(spec.id.clone()).encode(),
+    );
+
+    let s3_manager: &s3::Manager = s3_manager.as_ref();
+    let mut objects: Vec<aws_sdk_s3::model::Object>;
+    loop {
+        sleep(Duration::from_secs(20)).await;
+
+        objects =
+            s3::spawn_list_objects(s3_manager.clone(), s3_bucket, Some(s3_key_prefix.clone()))
+                .await
+                .map_err(|e| {
+                    Error::new(ErrorKind::Other, format!("failed spawn_list_objects {}", e))
+                })?;
+
+        log::info!(
+            "{} anchor nodes are ready (expecting {} nodes)",
+            objects.len(),
+            target_nodes
+        );
+
+        if objects.len() as u32 >= target_nodes {
+            break;
+        }
+    }
+
+    let mut s3_keys: Vec<String> = Vec::new();
+    for obj in objects.iter() {
+        let s3_key = obj.key().expect("unexpected None s3 object");
+        s3_keys.push(s3_key.to_string());
+    }
+    // s3 API should return the sorted list
+    // in the descending order of "last_modified" timestamps
+    // but to make sure, sort in lexicographical order
+    s3_keys.sort();
+
+    let mut bootstrap_ids: Vec<String> = vec![];
+    let mut bootstrap_ips: Vec<String> = vec![];
+    for s3_key in s3_keys.iter() {
+        // just parse the s3 key name
+        // to reduce "s3_manager.get_object" call volume
+        let ready_anchor_node = avalancheup_aws::StorageNamespace::parse_node_from_path(s3_key)?;
+
+        bootstrap_ids.push(ready_anchor_node.node_id);
+
+        // assume all nodes in the network use the same ports
+        // ref. "avalanchego/config.StakingPortKey" default value is "9651"
+        let staking_port = spec.avalanchego_config.staking_port;
+        bootstrap_ips.push(format!("{}:{}", ready_anchor_node.public_ip, staking_port));
+    }
+    log::info!("found {} seed nodes that are ready", bootstrap_ids.len());
+
+    Ok((bootstrap_ids, bootstrap_ips))
+}
+
 fn start_avalanche_systemd_service(
     avalanche_bin_path: &str,
-    avalanchego_config: avalanchego::config::Config,
+    avalanchego_config: &avalanchego::config::Config,
+    coreth_config: &coreth::config::Config,
 ) -> io::Result<()> {
     log::info!("STEP: setting up and starting Avalanche systemd service...");
 
+    fs::create_dir_all(&avalanchego_config.log_dir)?;
+    if let Some(v) = &avalanchego_config.subnet_config_dir {
+        fs::create_dir_all(Path::new(v).join("C"))?;
+    }
+    if let Some(v) = &avalanchego_config.profile_dir {
+        fs::create_dir_all(v)?;
+    }
+    if let Some(v) = &coreth_config.continuous_profiler_dir {
+        fs::create_dir_all(v)?;
+    }
+
+    avalanchego_config.validate()?;
+    if avalanchego_config.config_file.is_none() {
+        return Err(Error::new(
+            ErrorKind::InvalidInput,
+            "avalanchego_config.config_file is None",
+        ));
+    }
+
     // persist before starting the service
-    avalanchego_config
-        .sync(None)
-        .expect("failed to sync avalanchego config_file");
+    avalanchego_config.sync(None)?;
 
     // don't use "Type=notify"
     // as "avalanchego" currently does not do anything specific to systemd
@@ -1189,6 +1146,104 @@ async fn check_liveness(ep: &str) -> io::Result<()> {
     }
 
     Ok(())
+}
+
+/// if run in anchor nodes, the uploaded file will be downloaded
+/// in bootstrapping non-anchor nodes for custom networks
+async fn publish_node_info_ready_loop(
+    ec2_instance_id: Arc<String>,
+    node_kind: node::Kind,
+    node_id: Arc<String>,
+    public_ipv4: Arc<String>,
+    s3_manager: Arc<s3::Manager>,
+    s3_bucket: Arc<String>,
+    avalancheup_spec_path: Arc<String>,
+) {
+    log::info!("STEP: publishing node info for its readiness...");
+
+    let spec =
+        avalancheup_aws::Spec::load(avalancheup_spec_path.as_str()).expect("failed to load spec");
+
+    let http_scheme = {
+        if spec.avalanchego_config.http_tls_enabled.is_some()
+            && spec
+                .avalanchego_config
+                .http_tls_enabled
+                .expect("unexpected None avalanchego_config.http_tls_enabled")
+        {
+            "https"
+        } else {
+            "http"
+        }
+    };
+    let local_node = avalancheup_aws::Node::new(
+        node_kind.clone(),
+        ec2_instance_id.as_str(),
+        node_id.as_str(),
+        public_ipv4.as_str(),
+        http_scheme,
+        spec.avalanchego_config.http_port,
+    );
+    let node_info = avalancheup_aws::NodeInfo::new(
+        local_node.clone(),
+        spec.avalanchego_config.clone(),
+        spec.coreth_config.clone(),
+    );
+
+    let node_info_path = random_manager::tmp_path(10, Some(".yaml")).unwrap();
+    node_info.sync(node_info_path.clone()).unwrap();
+
+    let node_info_ready_s3_key = {
+        if matches!(node_kind, node::Kind::Anchor) {
+            avalancheup_aws::StorageNamespace::DiscoverReadyAnchorNode(
+                spec.id.clone(),
+                local_node.clone(),
+            )
+            .encode()
+        } else {
+            avalancheup_aws::StorageNamespace::DiscoverReadyNonAnchorNode(
+                spec.id.clone(),
+                local_node.clone(),
+            )
+            .encode()
+        }
+    };
+
+    let s3_manager: &s3::Manager = s3_manager.as_ref();
+    loop {
+        log::info!(
+            "STEP: posting node info ready for {}",
+            node_info.local_node.kind
+        );
+
+        match s3::spawn_put_object(
+            s3_manager.clone(),
+            &node_info_path,
+            &s3_bucket.to_string(),
+            &node_info_ready_s3_key,
+        )
+        .await
+        {
+            Ok(_) => {}
+            Err(e) => {
+                log::warn!("failed spawn_put_object {}", e);
+                sleep(Duration::from_secs(5)).await;
+                continue;
+            }
+        }
+
+        match fs::remove_file(&node_info_path) {
+            Ok(_) => {}
+            Err(e) => {
+                log::warn!("failed to remove_file {}", e);
+                sleep(Duration::from_secs(5)).await;
+                continue;
+            }
+        }
+
+        log::info!("sleeping 10-min for next 'publish_node_info_ready_loop'");
+        sleep(Duration::from_secs(600)).await;
+    }
 }
 
 /// returns "hello" from "a/b/c/hello.zstd"
