+++ bin/avalanched-aws/src/flags.rs
@@ -1,4 +1,5 @@
 /// Defines flag options.
 pub struct Options {
     pub log_level: String,
+    pub lite_mode: bool,
 }
